OK, a bunch of issues found...
Creating a Font from a BitmapFont produces very, very different vertical positioning relative to creating one from a .fnt (or using KnownFonts). It seems about half a line too low, but this may vary between fonts. Changing this would break a lot of UIs, so I might have to leave it as-is.

The offsets that can now be used with emoji and other atlases don't behave the same way if the atlas is added before scaling or after scaling... This one may be more solvable.