Add raw .fnt/.png files for Gentium Un Italic.
The idea here is to use Gentium Italic's many variations on its serifs and generally different "flow," but to undo the slant or oblique effect it has in Italic mode. This concept was Siavash Ranbar's idea, and he implemented it with a bold font at about the same time I had this version working with a regular-weight font. I did a few extra steps later, like restoring normal slant to some special characters like the multi-level parentheses used in math.

KnownFonts still needs to be updated.