The isMono flag was calculated differently!
It considered all glyphs in the constructor that takes a BitmapFont, but omitted the problematic full-block glyph with id 9608 in the constructor that takes a .fnt file. This probably made the BitmapFont-based constructors look no different much of the time, because all glyphs but 9608 had their normal width, but would change their horizontal position.