Fix isMono detection for at least some fonts.
Inconsolata LGC (standard) is absolutely monospaced, but wasn't detecting as such. This was because the minWidth didn't take into account the widthAdjust applied to glyphs. Another possible bug that fonts might have encountered was that having an empty kernings block wasn't equivalent to omitting a kernings block entirely, and a monospaced font with an empty kernings block wouldn't have registered as mono.