Just needed to reset size when we reset!
I'm not sure what this might affect if wrap is false, and other tests may show this needs to be only done when wrap is true. I think it is likely that setting the size to the initial value for a new TypingLabel is what most users expect, though (including me).