Going crazy here. Trying nested Tables.
This approach works in DetritusScion, and I have no idea what changed here. The background doesn't show up for some reason, and the location is all wrong... The root Table should have a centered nested Table in it just off the bottom of the screen, with that nested Table holding left-aligned TypingLabel items.